llenar tablas y empezando ha realizar el excel para subir prodcutos

+++ AVT_TRAKING-b5a19f52c2a889d8592336db032ecf143043f8f5/AVT_TRAKING/AVT_TRAKING/Documents/Creacion de DB y de la tabla de Usuarios.sql
@@ -391,12 +391,12 @@ GO
 
 create table rental(
 	type varchar(15)primary key not null,
-	leg float,
+	leg money,
 	plk float,
-	deck float,
-	ladder float,
+	deck money,
+	ladder money,
 	truckLoad float,
-	truck float
+	truck money
 )
 
 --##########################################################################################
@@ -1176,12 +1176,12 @@ GO
 
 --create table rental(
 --	type varchar(15)primary key not null,
---	leg float,
+--	leg money,
 --	plk float,
---	deck float,
---	ladder float,
+--	deck money,
+--	ladder money,
 --	truckLoad float,
---	truck float
+--	truck money
 --)
 
 --create table subJobs(
